Created conflateddict module
