+++ .gitignore
@@ -2,3 +2,7 @@
 __pycache__
 *pyc
 .cache
+.vscode
+.tox
+.pytest_cache
+*.egg-info
